Use different names for accounts so you can run the stress test multiple times without having to restart the validator nodes

+++ test/integration/integration_test.go
@@ -13,7 +13,7 @@ func TestExternalTestSuite(t *testing.T) {
 	}
 	t.Log(">>> Setting up connection to local node <<<")
 
-	seed := int64(testCommon.LookupEnvInt(t, "SEED", 0))
+	seed := testCommon.LookupEnvInt(t, "SEED", 0)
 	rpcMode := testCommon.LookupRpcMode(t, "RPC_MODE", testCommon.SingleRpc)
 	rpcEndpoints := testCommon.LookupEnvStringArray("RPC_URLS", []string{"http://localhost:26657"})
